pyproject.toml
[tool.poetry]
name = "scrape-and-notify"
version = "0.1.0"
description = "A web scraper with notifications"
authors = ["Kyler Frazier <kyfrazier100@gmail.com>"]
readme = "README.md"
packages = [{include = "scrape_and_notify", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.4"
beautifulsoup4 = "^4.13.4"
lxml = "^5.4.0"
discord-py = "^2.5.2"
aiohttp = "^3.12.13"

[tool.poetry.scripts]
scrape-and-notify = "scrape_and_notify.main:run"

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
target-version = "py310"

# Exclude common directories
exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    "__pycache__",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "W"]
ignore = []
fixable = ["ALL"]
